+++ src/trades_call_builder.ts
@@ -55,6 +55,16 @@ export class TradesCallBuilder extends CallBuilder<
   }
 
   /**
+   * Filter trades by a specific type.
+   * @param {ServerApi.TradeType} tradeType the trade type to filter by.
+   * @returns {TradesCallBuilder} current TradesCallBuilder instance.
+   */
+  public forType(tradeType: ServerApi.TradeType): this {
+    this.url.setQuery("trade_type", tradeType);
+    return this;
+  }
+
+  /**
    * Filter trades for a specific account
    * @see [Trades for Account](https://www.stellar.org/developers/horizon/reference/endpoints/trades-for-account.html)
    * @param {string} accountId For example: `GBYTR4MC5JAX4ALGUBJD7EIKZVM7CUGWKXIUJMRSMK573XH2O7VAK3SR`